Add order on group by
